+++ apps/api-journeys-modern/src/emails/stories/JourneyShared.stories.tsx
@@ -1,6 +1,7 @@
 import { Meta, StoryObj } from '@storybook/react'
 
-import { JourneyWithTeamAndUserJourney } from '../../types'
+import { JourneyWithTeamAndUserJourney } from '@core/yoga/email/types'
+
 import { JourneySharedEmail } from '../templates/JourneyShared'
 
 import { apiJourneysConfig } from './apiJourneysConfig'
