+++ apps/api-journeys-modern/src/emails/templates/JourneyShared/JourneySharedNoAccount/JourneySharedNoAccount.tsx
@@ -20,10 +20,11 @@ import {
   Header,
   UnsubscribeLink
 } from '@core/yoga/email/components'
-import { JourneyForEmails } from '@core/yoga/email/types/types'
+import {
+  JourneyForEmails,
+  JourneyWithTeamAndUserJourney
+} from '@core/yoga/email/types'
 import { User } from '@core/yoga/firebaseClient'
-
-import { JourneyWithTeamAndUserJourney } from '../../../../types'
 
 interface JourneySharedNoAccountEmailProps {
   journey: JourneyWithTeamAndUserJourney
