+++ apps/api-journeys-modern/src/workers/email/service/service.spec.ts
@@ -8,18 +8,18 @@ import {
   UserTeamRole
 } from '.prisma/api-journeys-modern-client'
 import { sendEmail } from '@core/yoga/email'
-
-import { prismaMock } from '../../../../test/prismaMock'
-
 import {
   JourneyAccessRequest,
   JourneyEditInviteJob,
   JourneyRequestApproved,
   TeamInviteAccepted,
   TeamInviteJob,
-  TeamRemoved,
-  service
-} from './service'
+  TeamRemoved
+} from '@core/yoga/email/types'
+
+import { prismaMock } from '../../../../test/prismaMock'
+
+import { service } from './service'
 
 jest.mock('@apollo/client')
