+++ apps/api-journeys/src/app/app.module.ts
@@ -14,7 +14,6 @@ import { LoggerModule } from 'nestjs-pino'
 import { ActionModule } from './modules/action/action.module'
 import { BlockModule } from './modules/block/block.module'
 import { CustomDomainModule } from './modules/customDomain/customDomain.module'
-import { EmailModule } from './modules/email/email.module'
 import { EventModule } from './modules/event/event.module'
 import { NestHealthModule } from './modules/health/health.module'
 import { HostModule } from './modules/host/host.module'
@@ -40,7 +39,6 @@ import { VisitorModule } from './modules/visitor/visitor.module'
     ActionModule,
     BlockModule,
     CustomDomainModule,
-    EmailModule,
     EventModule,
     HostModule,
     IntegrationModule,
