+++ apps/api-journeys/src/app/modules/userInvite/userInvite.service.spec.ts
@@ -1,7 +1,8 @@
 import { getQueueToken } from '@nestjs/bullmq'
 import { Test, TestingModule } from '@nestjs/testing'
 
-import { JourneyWithTeamAndUserJourney } from '../email/email.consumer'
+import { JourneyWithTeamAndUserJourney } from '@core/yoga/email/types'
+
 import { UserInviteModule } from '../userInvite/userInvite.module'
 
 import { UserInviteService } from './userInvite.service'
